fix: correct import names due to incomplete refactoring
