+++ tests/services/data_manager/data_loaders/test_load_csv.py
@@ -1,4 +1,4 @@
-from services.data_loaders import load_csv
+from services.data_manager.data_loaders import load_csv
 from pathlib import Path
