+++ tests/services/data_manager/data_loaders/test_load_excel.py
@@ -1,4 +1,4 @@
-from services.data_loaders import load_excel
+from services.data_manager.data_loaders import load_excel
 from pathlib import Path
